# api/openapi.yaml
openapi: 3.0.3
info:
  title: Антискам тренажёр API
  version: 1.0.0
  description: |
    REST API интерактивного тренажёра безопасных сделок (Кейс 5).

    Спецификация — **единственный источник истины** контракта между фронтендом
    и бэкендом (MNT4). Любое изменение поведения API начинается с изменения
    этого файла, а не кода.

    ## Соглашения

    - Базовый префикс всех прикладных ресурсов — `/api/v1`.
    - Тела запросов и ответов — `application/json; charset=utf-8`.
    - Отметки времени — RFC 3339 в UTC (`2026-08-03T12:30:00Z`).
    - Идентификаторы сессий — UUID v4. Сценарии, шаги, варианты и признаки
      риска адресуются стабильными строковыми кодами, а не числовыми id.
    - Любая ошибка возвращается в едином конверте (`Error`) с машинным кодом
      из перечисления `ErrorCode`. Внутреннее устройство системы в ошибках не
      раскрывается (SEC6).
    - Все входные данные проверяются на сервере (SEC3): ограничения, описанные
      в схемах, обязательны к реализации на бэкенде и не считаются выполненными
      за счёт валидации на клиенте.

    ## Аутентификация

    JWT в заголовке `Authorization: Bearer <token>`. Токен выдаётся при
    регистрации и входе (FR1, FR2).

    Эндпоинты делятся на три группы:

    | Группа | Поведение |
    |---|---|
    | Публичные | Токен не требуется: витрина сценариев, справочник (FR4) |
    | С опциональным токеном | Работают без токена, но с ним ответ дополняется прогрессом пользователя |
...

...

...

...

  contact:
...
  license:
...
servers:
...
paths:
...
  /api/v1/scenarios:
    get:
      tags: [catalog]
      operationId: listScenarios
      summary: Витрина сценариев
      description: |
        Возвращает сценарии, доступные для выбранной роли (FR5, FR6).
        Набор небольшой и выводится целиком: фильтрации, поиска и пагинации нет.

        Для авторизованного пользователя карточка дополняется полями прогресса:
        `attempted`, `best_percent`, `best_level`, `last_attempt_at` (FR6, FR7).
        Для гостя эти поля отсутствуют (FR4).
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: role
          in: query
          required: false
          description: Роль пользователя. Без параметра возвращаются сценарии обеих ролей.
          schema:
            $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: Список сценариев
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/scenarios/{scenarioCode}:
    get:
      tags: [catalog]
      operationId: getScenario
      summary: Карточка сценария перед стартом
      description: |
        Данные экрана подтверждения старта: описание ситуации, роль, сложность,
        ожидаемое число шагов. Шаги и варианты ответа здесь не отдаются — они
        приходят по одному в ходе прохождения.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ScenarioCode"
      responses:
        "200":
          description: Карточка сценария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/sessions:
    post:
      tags: [training]
      operationId: startSession
      summary: Старт тренировки
      description: |
        Создаёт сессию прохождения и возвращает первый шаг (FR8).
        Сессия жёстко связывается с текущей версией сценария (FR32):
        последующая правка контента не меняет результат этой попытки.

        Если по сценарию уже есть незавершённая сессия, возвращается
        `409 session_already_active` с её идентификатором в `details.session_id`
        — фронтенд предлагает продолжить или начать заново (прецедент 2, п. 6б).
        Повторный запрос с `restart: true` помечает старую сессию прерванной и
        создаёт новую (прецедент 8, п. 2а).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionRequest"
      responses:
        "201":
          description: Сессия создана, отдан первый шаг
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: По сценарию есть незавершённая сессия (`session_already_active`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: session_already_active
                  message: По этому сценарию есть незавершённая тренировка
                  details:
                    session_id: 3f1c9c1e-6f2a-4f7f-9a7b-6b5f2e2f1a11
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}:
    get:
      tags: [training]
      operationId: getSession
      summary: Состояние сессии
      description: |
        Текущий шаг, позиция в сценарии и накопленный балл. Используется для
        восстановления экрана прохождения после перезагрузки страницы и для
        продолжения прерванной тренировки (FR12, REL2).

        Чужая сессия отдаётся как `404` (SEC2).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Состояние сессии
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/answers:
    post:
      tags: [training]
      operationId: submitAnswer
      summary: Зафиксировать выбор на шаге
      description: |
        Фиксирует выбор пользователя, начисляет баллы и возвращает обучающую
        обратную связь вместе со следующим шагом (FR10, FR11, FR15–FR17, FR20).

        Ответ содержит всё, что нужно экрану прохождения, — повторный запрос
        состояния сессии не требуется (USR3).

        **Идемпотентность (FR13).** Если шаг уже отвечен, возвращается ранее
        зафиксированный результат с `already_answered: true`; состояние сессии
        не меняется и баллы повторно не начисляются. Это верно и в случае,
        когда прислан другой `option_code`: изменить сделанный выбор нельзя.

        **Терминальный шаг (FR14).** Если выбранный вариант ведёт в терминальный
        шаг, сессия завершается: `session_finished: true`, `session.status`
        становится `completed`, разбор доступен по `/sessions/{id}/result`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerRequest"
      responses:
        "200":
          description: Выбор зафиксирован (или уже был зафиксирован ранее)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
        "400":
          description: |
            Ошибка валидации. Возможные коды:
            `validation_error` — тело запроса не соответствует схеме;
            `step_not_current` — шаг не является текущим шагом сессии;
            `option_not_found` — варианта с таким кодом нет на этом шаге
            (прецедент 3, п. 4б).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: option_not_found
                  message: Указанного варианта нет на этом шаге
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Сессия уже завершена или прервана (`session_finished`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: session_finished
                  message: Тренировка уже завершена
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/result:
    get:
      tags: [training]
      operationId: getSessionResult
      summary: Итоговый разбор прохождения
      description: |
        Полный разбор завершённой сессии (прецедент 4): балл и уровень
        безопасности (FR20, FR21), сравнение с предыдущей попыткой (FR23),
        пошаговый разбор с безопасными альтернативами (FR18), карта
        распознанных и пропущенных признаков риска, практические рекомендации
        (FR19) и предложение следующего шага обучения (FR27).

        Разбор воспроизводим: одна и та же последовательность выборов на одной
        версии сценария всегда даёт один и тот же результат (FR22).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Разбор прохождения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Сессия ещё не завершена (`session_not_finished`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: session_not_finished
                  message: Тренировка ещё не завершена
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/abandon:
    post:
      tags: [training]
      operationId: abandonSession
      summary: Прервать тренировку
      description: |
        Помечает незавершённую сессию прерванной (`abandoned`). Используется,
        когда пользователь выбирает «начать заново» (прецедент 8, п. 2а).
        Прерванная сессия не участвует в прогрессе и не имеет разбора.
        Операция идемпотентна: повторный вызов на уже прерванной сессии — `204`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204":
          description: Сессия прервана
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Сессия уже завершена, прервать нельзя (`session_finished`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # ───────────────────────────── progress ─────────────────────────────
...
  /api/v1/risk-signals:
    get:
      tags: [reference]
      operationId: listRiskSignals
      summary: Справочник признаков риска
      description: |
        Признаки мошенничества с описанием и рекомендацией по безопасному
        поведению (FR28). Доступен без авторизации (FR4).
      security: []
      parameters:
        - name: side
          in: query
          required: false
          description: Сторона сделки. Без параметра возвращаются признаки обеих сторон.
          schema:
            $ref: "#/components/schemas/Side"
      responses:
        "200":
          description: Список признаков риска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskSignalListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/risk-signals/{signalCode}:
    get:
      tags: [reference]
      operationId: getRiskSignal
      summary: Карточка признака риска
      description: |
        Открывается по ссылке из разбора прохождения (FR29) и из справочника.
        Дополнительно перечисляет сценарии, в которых признак отрабатывается
        (прецедент 9, п. 5).
      security: []
      parameters:
        - $ref: "#/components/parameters/SignalCode"
      responses:
        "200":
          description: Карточка признака риска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskSignalDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT, выданный `/api/v1/auth/register` или `/api/v1/auth/login`.
  parameters:
    ScenarioCode:
      name: scenarioCode
      in: path
      required: true
      description: Стабильный код сценария.
      schema:
        $ref: "#/components/schemas/ScenarioCode"
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Идентификатор сессии прохождения.
      schema:
        type: string
        format: uuid
      example: 3f1c9c1e-6f2a-4f7f-9a7b-6b5f2e2f1a11
    SignalCode:
      name: signalCode
      in: path
      required: true
      description: Код признака риска из каталога.
      schema:
        $ref: "#/components/schemas/RiskSignalCode"
  responses:
    ValidationError:
      description: Данные запроса не прошли проверку на сервере (`validation_error`, SEC3)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: validation_error
              message: Проверьте правильность заполнения полей
              details:
                fields:
                  nickname: Допустимы латинские буквы, цифры, дефис и подчёркивание
    Unauthorized:
      description: Требуется авторизация (`unauthorized`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: Требуется вход в систему
    NotFound:
      description: |
        Ресурс не найден (`not_found`). Тот же ответ отдаётся, если ресурс
        существует, но принадлежит другому пользователю (SEC2).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: Ресурс не найден
    InternalError:
      description: |
        Внутренняя ошибка (`internal_error`). Подробности не раскрываются;
        для диагностики используется `request_id` из ответа и логов (SEC6, MNT6).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal_error
              message: Внутренняя ошибка сервиса
            request_id: 01J8XQ2M7K3ZC0P4B6R9V2T5
  schemas:
...
    Role:
      type: string
      description: Роль пользователя в сделке.
      enum: [buyer, seller]
      example: buyer
    Side:
      type: string
      description: Сторона сделки, для которой актуален признак риска.
      enum: [buyer, seller]
      example: buyer
...
    SessionStatus:
      type: string
      description: |
        Состояние сессии: `in_progress` — идёт, `completed` — завершена
        и имеет разбор, `abandoned` — прервана пользователем, в прогрессе
        не учитывается.
      enum: [in_progress, completed, abandoned]
      example: in_progress
...
    ScenarioCode:
      type: string
      description: Стабильный код сценария, не меняется между версиями.
      pattern: "^[a-z0-9-]{3,64}$"
      example: booking-link
    StepCode:
      type: string
      description: Код шага, уникальный в пределах сценария.
      pattern: "^[a-z0-9_-]{1,32}$"
      example: s3
...
    ErrorCode:
      type: string
      description: |
        Машинный код ошибки. Фронтенд принимает решения по коду, а не по
        тексту `message` — тексты могут меняться.
      enum:
        - validation_error
        - unauthorized
        - invalid_credentials
        - forbidden
        - not_found
        - nickname_taken
        - session_already_active
        - session_finished
        - session_not_finished
        - step_not_current
        - option_not_found
        - internal_error
      example: validation_error
    Error:
      type: object
      description: Единый конверт ошибки. Внутреннее устройство системы не раскрывается (SEC6).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Человекочитаемое сообщение на русском языке, пригодное для показа пользователю.
              example: Проверьте правильность заполнения полей
            details:
              type: object
              additionalProperties: true
              description: |
                Дополнительный контекст. Для `validation_error` — объект
                `fields` с сообщениями по полям; для `session_already_active` —
                `session_id` незавершённой сессии.
        request_id:
          type: string
          description: Идентификатор запроса для сопоставления с логами (MNT6).
          example: 01J8XQ2M7K3ZC0P4B6R9V2T5
...
    ScenarioRef:
      type: object
      description: Краткая ссылка на сценарий, встраиваемая в другие ответы.
      required: [code, title, role, difficulty, version]
      properties:
        code:
          $ref: "#/components/schemas/ScenarioCode"
        title:
          type: string
          example: Бронь по ссылке
        role:
          $ref: "#/components/schemas/Role"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        version:
          type: integer
          minimum: 1
          description: Версия контента сценария (FR32).
          example: 1
    ScenarioCard:
      allOf:
        - $ref: "#/components/schemas/ScenarioRef"
        - type: object
          required: [description, steps_count]
          properties:
            description:
              type: string
              description: Краткое описание ситуации для карточки на витрине.
              example: Продавец просит внести предоплату по ссылке, чтобы «снять объявление с публикации».
            steps_count:
              type: integer
              minimum: 3
              maximum: 8
              description: |
                Ожидаемое число шагов. Из-за ветвления фактическая длина пути
                может отличаться на шаг; значение используется для индикатора
                прогресса (USR2).
              example: 6
            risk_signal_codes:
              type: array
              description: Признаки риска, отрабатываемые сценарием.
              items:
                $ref: "#/components/schemas/RiskSignalCode"
              example: [RISK_PHISHING_LINK, RISK_PREPAY_OUTSIDE, RISK_URGENCY_PRESSURE]
            attempted:
              type: boolean
              description: Проходился ли сценарий ранее (FR6). Только для авторизованного пользователя.
              example: true
            best_percent:
              allOf:
                - $ref: "#/components/schemas/Percent"
              nullable: true
              description: Лучший результат пользователя по сценарию (FR7). `null`, если завершённых попыток нет.
            best_level:
              allOf:
                - $ref: "#/components/schemas/SecurityLevel"
              nullable: true
              description: Уровень безопасности лучшей попытки.
            attempts_count:
              type: integer
              minimum: 0
              description: Число завершённых попыток пользователя по сценарию.
              example: 2
            last_attempt_at:
              type: string
              format: date-time
              nullable: true
              description: Дата последней завершённой попытки.
              example: "2026-08-02T18:40:00Z"
    ScenarioDetail:
      allOf:
        - $ref: "#/components/schemas/ScenarioCard"
        - type: object
          properties:
            intro:
              type: string
              description: "Вводный текст экрана подтверждения старта: контекст сделки и что предстоит сделать."
              example: Вы покупатель. Нашли нужный товар по адекватной цене и написали продавцу.
            estimated_minutes:
              type: integer
              minimum: 1
              description: Ориентировочное время прохождения в минутах.
              example: 3
            active_session_id:
              type: string
              format: uuid
              nullable: true
              description: Незавершённая сессия пользователя по этому сценарию, если она есть (FR12).
    ScenarioListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioCard"

    # ───────────────────────── прохождение ────────────────────────────
    StartSessionRequest:
      type: object
      required: [scenario_code]
      properties:
        scenario_code:
          $ref: "#/components/schemas/ScenarioCode"
        restart:
          type: boolean
          default: false
          description: |
            Начать заново, прервав незавершённую сессию по этому сценарию.
            Без флага такая ситуация даёт `409 session_already_active`.
...
    StepView:
      type: object
      description: |
        Шаг, каким его видит пользователь. Признаки риска шага здесь не
        передаются — они раскрываются в обратной связи после выбора (FR16).
      required: [code, type, position, content, options]
      properties:
        code:
          $ref: "#/components/schemas/StepCode"
        type:
          $ref: "#/components/schemas/StepType"
        position:
          type: integer
          minimum: 1
          description: Порядковый номер шага на пройденном пути, с 1 (USR2).
          example: 3
        content:
          $ref: "#/components/schemas/StepContent"
        options:
          type: array
          description: |
            Ровно три варианта для шага типа `dialog`; пустой массив для
            терминального шага. Порядок вариантов задаётся контентом и не
            коррелирует с их безопасностью.
          maxItems: 3
          items:
            $ref: "#/components/schemas/OptionView"
    SessionState:
      type: object
      description: Состояние сессии и текущий шаг — всё, что нужно экрану прохождения.
      required:
        [id, scenario, status, score, answers_count, steps_total, started_at]
      properties:
        id:
          type: string
          format: uuid
          example: 3f1c9c1e-6f2a-4f7f-9a7b-6b5f2e2f1a11
        scenario:
          $ref: "#/components/schemas/ScenarioRef"
        status:
          $ref: "#/components/schemas/SessionStatus"
        current_step:
          allOf:
            - $ref: "#/components/schemas/StepView"
          nullable: true
          description: Текущий шаг. `null`, если сессия завершена или прервана.
        score:
          type: integer
          description: Накопленный балл — сумма весов сделанных выборов (FR20).
          example: 20
        answers_count:
          type: integer
          minimum: 0
          description: Число зафиксированных выборов.
          example: 2
        steps_total:
          type: integer
          description: Ожидаемое число шагов сценария, для индикатора «шаг X из Y» (USR2).
          example: 6
        started_at:
          type: string
          format: date-time
          example: "2026-08-03T12:00:00Z"
        finished_at:
          type: string
          format: date-time
          nullable: true
          example: null
    SubmitAnswerRequest:
      type: object
      description: |
        `step_code` передаётся явно и проверяется на сервере: это защищает от
        рассинхронизации клиента с состоянием сессии (например, после
        повторной отправки формы или кнопки «назад» в браузере).
      required: [step_code, option_code]
      properties:
        step_code:
          $ref: "#/components/schemas/StepCode"
        option_code:
          $ref: "#/components/schemas/OptionCode"
...
    AnswerResult:
      type: object
      description: |
        Обучающая обратная связь и следующий шаг — один ответ на весь цикл
        «выбрал → понял → пошёл дальше» (USR3).
      required:
        [
          step_code,
          option_code,
          outcome,
          score_delta,
          feedback,
          risk_signals,
          already_answered,
          session_finished,
          session,
        ]
      properties:
        step_code:
          $ref: "#/components/schemas/StepCode"
        option_code:
          $ref: "#/components/schemas/OptionCode"
        outcome:
          $ref: "#/components/schemas/Outcome"
        score_delta:
          type: integer
          description: "Вес выбранного варианта: `+10`, `0` или `−10` (FR20)."
          enum: [10, 0, -10]
          example: -10
        feedback:
          type: string
          description: |
            Объяснение по схеме «признак → причина → альтернатива» (USR4, FR15).
            Формулировки «правильно» и «неправильно» контентом не используются.
          example: >-
            Домен только похож на официальный: настоящие страницы оплаты живут
            на avito.ru. Оплата по ссылке из переписки не защищена — деньги
            уходят напрямую мошеннику. Безопасный путь — кнопка оплаты внутри
            объявления.
        risk_signals:
          type: array
          description: Признаки риска, присутствовавшие на этом шаге (FR16).
          items:
            $ref: "#/components/schemas/RiskSignalRef"
        safe_alternative:
          allOf:
            - $ref: "#/components/schemas/SafeAlternative"
          nullable: true
        already_answered:
          type: boolean
          description: |
            `true`, если шаг был отвечен ранее и возвращён сохранённый
            результат: состояние не изменилось, баллы не начислены (FR13).
          example: false
        session_finished:
          type: boolean
          description: Сессия завершена достижением терминального шага (FR14).
          example: false
        session:
          allOf:
            - $ref: "#/components/schemas/SessionState"
          description: Состояние сессии после фиксации выбора, с уже подставленным следующим шагом.

    # ───────────────────────── результат ──────────────────────────────
...
    SessionResult:
      type: object
      description: Экран результата целиком (прецедент 4).
      required:
        [session_id, scenario, score, breakdown, signals, recommendations, next_step, started_at, finished_at]
      properties:
        session_id:
          type: string
          format: uuid
          example: 3f1c9c1e-6f2a-4f7f-9a7b-6b5f2e2f1a11
        scenario:
          $ref: "#/components/schemas/ScenarioRef"
        score:
          $ref: "#/components/schemas/ScoreSummary"
        comparison:
          allOf:
            - $ref: "#/components/schemas/AttemptComparison"
          nullable: true
        breakdown:
          type: array
          description: Шаги в порядке прохождения (FR18).
          items:
            $ref: "#/components/schemas/BreakdownItem"
        signals:
          type: array
          description: Признаки риска сценария с отметкой, распознал их пользователь или пропустил.
          items:
            $ref: "#/components/schemas/SessionSignalOutcome"
        recommendations:
          type: array
          description: Практические рекомендации, применимые к реальной сделке (FR19).
          items:
            type: string
          example:
            - Оплачивайте только внутри платформы — ссылки из переписки не защищены.
            - Просьба уйти в мессенджер — сама по себе признак риска, независимо от объяснения.
        next_step:
          $ref: "#/components/schemas/NextStepSuggestion"
        started_at:
          type: string
          format: date-time
          example: "2026-08-03T12:00:00Z"
        finished_at:
          type: string
          format: date-time
          example: "2026-08-03T12:04:30Z"
        duration_seconds:
          type: integer
          minimum: 0
          example: 270

    # ────────────────────────── прогресс ──────────────────────────────
...
    RiskSignalDetail:
      allOf:
        - $ref: "#/components/schemas/RiskSignal"
        - type: object
          properties:
            related_scenarios:
              type: array
              description: Сценарии, в которых признак отрабатывается (прецедент 9, п. 5).
              items:
                $ref: "#/components/schemas/ScenarioRef"
    RiskSignalListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RiskSignal"